requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
requests
beautifulsoup4
python-dateutil
APScheduler
